[#147] Renamed the references to `stalker.models.link` module to `stalker.models.file`.
