+++ docs/source/inheritance_diagram.rst
@@ -28,8 +28,8 @@ Inheritance Diagram
    stalker.models.entity.Entity
    stalker.models.entity.EntityGroup
    stalker.models.entity.SimpleEntity
+   stalker.models.file.File
    stalker.models.format.ImageFormat
-   stalker.models.link.Link
    stalker.models.message.Message
    stalker.models.mixins.ACLMixin
    stalker.models.mixins.CodeMixin
@@ -48,7 +48,7 @@ Inheritance Diagram
    stalker.models.repository.Repository
    stalker.models.review.Review
    stalker.models.review.Daily
-   stalker.models.review.DailyLink
+   stalker.models.review.DailyFile
    stalker.models.scene.Scene
    stalker.models.schedulers.SchedulerBase
    stalker.models.schedulers.TaskJugglerScheduler
@@ -69,6 +69,7 @@ Inheritance Diagram
    stalker.models.ticket.TicketLog
    stalker.models.type.EntityType
    stalker.models.type.Type
+   stalker.models.variant.Variant
    stalker.models.version.Version
    stalker.models.wiki.Page
    :parts: 1
